grid coordinate tracking
Grid area 0 value's are now overridden when the player enters the room. The 0 is replaced with the running room_count total and that value in the grid (in the next update) will be able to be referenced via the room_dict to find the type (How many doors) of room it is.