OKAY So here's the deal...
I think I know where to go to complete this project, but honestly, I need to move on. Not only did I take a HUGE leap in attempting this project, but I took it before doing the research into text-based adventure games that would be required for my game to be fun. I have learned quite a bit about dictionaries and I/O Operations through this project and I'm grateful for the experience. I think I will work on something similar to this in the future, but just not right now.